beta_bot/requirements.txt
python-telegram-bot version=13.15
requests=2.22.0
python version=3.11